models pra migration

+++ LivrariaSouza.Models/Models/RegistroDeVenda.cs
@@ -17,7 +17,7 @@ namespace LivrariaSouza.Models.Models
         public int LivroId { get; set; }
         public string Titulo { get; set; }
         public string CapaLivro { get; set; }
-        public int ValorUnit { get; set; }
+        public decimal ValorUnit { get; set; }
         public int Quantidade { get; set; }
     }
